Fix build break on linux (compiler warning).
TBR=evan

Review URL: http://codereview.chromium.org/28231

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10540 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/extensions/extension.h
@@ -65,7 +65,7 @@ class Extension {
   static const char* kInvalidPluginsDirError;
 
   // The number of bytes in a legal id.
-  static const int kIdSize;
+  static const size_t kIdSize;
 
   // Creates an absolute url to a resource inside an extension. The
   // |extension_url| argument should be the url() from an Extension object. The
